requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
# RAG with Gemini API
sentence-transformers==2.7.0
huggingface_hub==0.24.0
# Note: older faiss-cpu versions (e.g. 1.8.0) may be unavailable for newer Python runtimes on Render.
# Use a newer release that has prebuilt wheels (adjust if needed for your environment).
faiss-cpu==1.13.1
google-generativeai==0.3.0
python-multipart==0.0.6
rich==13.3.5
# Pin a numpy release that supports Python 3.13 runtimes on Render
# (1.26.0 does not support Python 3.13). Using 1.26.4 which has wheels available.
numpy==1.26.4
